+++ src/component_builder/config.py
@@ -1,0 +1,22 @@
+from ConfigParser import ConfigParser
+
+from .component import Component
+
+
+def read_component_configuration(builder_ini_file):
+    config = ConfigParser(defaults={
+        'downstream': '',
+        'release-process': ''
+    })
+    config.readfp(builder_ini_file)
+
+    for component in config.sections():
+        kwargs = {
+            'title': component,
+            'release_process': config.get(component, 'release-process'),
+            'path': config.get(component, 'path')
+        }
+        downstream = config.get(component, 'downstream')
+        if downstream:
+            kwargs['downstream'] = downstream.split(',')
+        Component(**kwargs)
